+++ src/main/java/com/remind/api/mood/dto/response/ActivityListResponseDto.java
@@ -1,5 +1,6 @@
 package com.remind.api.mood.dto.response;
 
+import com.remind.api.mood.dto.ActivityListDto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 @Schema(description = "추가한 활동 목록")
